Ensure Laboratory is thread-safe
Here at Butternut, we use a Puma web server for our core application.
This means we have to always be mindful of the impact we have on the
multi-threaded environment we work in. There is a great summary here of
some of the issues that can occur in a Puma environment if thread safety
isn't managed carefully:

https://www.rubytapas.com/2016/11/20/ruby-thread-local-variables/

We want to make sure Laboratory is thread safe. Upto now, the Config
definition in Laboratory was not threadsafe as it was a single global
class with attr_accessor variables that every thread would change. This
commit solves this by doing the following:

* Redesigns `Laboratory::Config` to move the attr_accessor's to the
instance of the class, rather than the class itself.
* Then redefines Laboratory.config to use the current thread's
`laboratory_config` variable, or defines that variable as a new instance
of `Laboratory::Config`.

This results in each thread having a separate config instance.