Refactor position of the adapter
In a previous commit, we made the `Laboratory::Config` class threadsafe.
However, by doing this, we made the adapter variable local to the
current thread. This would mean having to define it on every thread,
which is unlikely to be what is required. This commit moves it up a
level onto the `Laboratory` object so it can be accessed directly from
there as a single global attribute.